+++ src/misc/macros.h
@@ -26,4 +26,8 @@
 #define HANDLE_RIGHT  0x08
 #define HANDLE_ABS    0x10
 
+#define LAYER_BELOW   kCGBackstopMenuLevelKey
+#define LAYER_NORMAL  kCGNormalWindowLevelKey
+#define LAYER_ABOVE   kCGFloatingWindowLevelKey
+
 #endif